+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to the "find-more" extension will be documented in this file
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
-- Initial release
 
+## [0.0.2] 2017-04-30
+### Added
+- Show find results in symbol reference panel.
+- Show find results in problems panel.
+
+### Fixed
+- Only the first one result can be found. 
+
+## [0.0.1] 2017-03
 ### Added
 - Title menu for find function to active "find all" in current file.